Reset everything back into main
